feat(alert): Alert Activities
